requirements.txt
fastmcp>=0.1.0
jinja2>=3.0.0
python-dotenv>=0.19.0